[test] Move utility functions to its own module
This keeps things cleaner and more organized.
