+++ z3c/dependencychecker/tests/test_main.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from z3c.dependencychecker.dependencychecker import main
+from z3c.dependencychecker.tests.utils import change_dir
 import mock
 import os
 import pkg_resources
@@ -7,23 +8,6 @@ import shutil
 import sys
 import tempfile
 import pytest
-
-
-class change_dir(object):
-    """Step into a directory temporarily
-
-    Copied from https://pythonadventures.wordpress.com/2013/12/15/chdir-a-context-manager-for-switching-working-directories/  # noqa: E501
-    """
-
-    def __init__(self, path):
-        self.old_dir = os.getcwd()
-        self.new_dir = path
-
-    def __enter__(self):
-        os.chdir(self.new_dir)
-
-    def __exit__(self, *args):
-        os.chdir(self.old_dir)
 
 
 @pytest.fixture(scope='module')
